requirements.txt
python-telegram-bot==21.6
apscheduler==3.10.4
pandas==2.2.3
openpyxl==3.1.5
ping3==4.0.8
python-dotenv==1.0.1
